Refactored collision detection, room objects into separate classes
