+++ Jamfile
@@ -151,6 +151,7 @@ COMMON_NAMES =
 	PathFont-font
 	DrawLines
 	ColorProgram
+	Collision
 	Scene
 	Mesh
 	load_save_png
